+++ back/src/main/java/com/openclassrooms/starterjwt/services/UserService.java
@@ -1,5 +1,8 @@
 package com.openclassrooms.starterjwt.services;
 
+import com.openclassrooms.starterjwt.exception.EmailExistingException;
+import com.openclassrooms.starterjwt.exception.NotFoundException;
+import com.openclassrooms.starterjwt.exception.UnauthorizedRequestException;
 import com.openclassrooms.starterjwt.models.User;
 import com.openclassrooms.starterjwt.payload.request.LoginRequest;
 import com.openclassrooms.starterjwt.payload.request.SignupRequest;
@@ -11,8 +14,11 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+
+import java.util.Objects;
 
 @Service
 public class UserService {
@@ -54,9 +60,9 @@ public class UserService {
                 isAdmin);
     }
 
-    public User create(SignupRequest signUpRequest) {
+    public void create(SignupRequest signUpRequest) {
         if (userRepository.existsByEmail(signUpRequest.getEmail())) {
-            return null;
+            throw new EmailExistingException();
         }
 
         // Create new user's account
@@ -66,14 +72,24 @@ public class UserService {
                 passwordEncoder.encode(signUpRequest.getPassword()),
                 false);
 
-        return userRepository.save(user);
+        userRepository.save(user);
     }
 
     public void delete(Long id) {
+        User user = this.findById(id);
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!Objects.equals(userDetails.getUsername(), user.getEmail())) {
+            throw new UnauthorizedRequestException();
+        }
         this.userRepository.deleteById(id);
     }
 
     public User findById(Long id) {
-        return this.userRepository.findById(id).orElse(null);
+        User user = this.userRepository.findById(id).orElse(null);
+        if (user == null) {
+            throw new NotFoundException();
+        }
+        return user;
     }
 }
